37 using ping in buckets (#40)
Full buckets now ping the oldest node (the back of the list) to see if it is awake.. If it is awake the 21 node is discarded, if it is down the old node is discarded and the new node is added to the front of the list. Testing is added.

Issue #39 (Receiving messages adds to routing table) was so small (like 3 lines), so I based this branch in it's branch and will add both to main at the same time.